Port to official Go SDK
Now that we have an official Go SDK [0], use it.

This neatly removes a lot of handwritten API call code. We don't get to
use the full power of the SDK because this tool renders its own JSON for
query and write requests, but that's its own cool validation of the
escape hatches that Stainless provides.

[0]: https://github.com/turbopuffer/turbopuffer-go